+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [v0.6.0]
+### Added
+- New method CheckStats on results package that validates AnsiblePlaybookJSONResults stats
+
+### Changed
+- __JSONStdoutCallbackResults__ on results package does not manipulates ansible JSON output, writes output as is into a writer
+- __JSONParser__ on results package has changed its signature to _JSONParse(data []byte) (*AnsiblePlaybookJSONResults, error)_
+- __simple-ansibleplaybook-json__ example has been modified to use a custom executor to manipulate the JSON output.
+- Use github.com/apenella/go-common-utils/error to manage errors
 
 ## [v0.5.1]
 ### Fixed